Add rpath support for even more dynamic linking
Previously, the user would have to manually compile the program because
the location of the library would vary from system to system. With
rpath, (most) systems will search for the library in the specified
directory you give during compile time.
If I did a poor job at explaining it, see
https://cmake.org/Wiki/CMake_RPATH_handling#What_is_RPATH_.3F